// src/app/model/ma-work-order.model.ts
export interface IMaWorkOrder {
    clicked?: boolean;
    accessCode: any;
    agreedDeliveryDate: any;
    amendedDate: any;
    approvedAt: any;
    approvedBy: any;
    automationSendAt: any;
    automationSendBy: any;
    automationStatus: any;
    bookingDate: any;
    bookingForexPrice: any;
    bookingPrice: any;
    bookingQuantity: any;
    brokerId: any;
    calculateType: any;
    capacity: any;
    certificationId: any;
    cmtForexId: any;
    cmtPrice: any;
    companyId: any;
    controlCode: any;
    countryId: any;
    costCenterId: any;
    createdByBooking: any;
    currentAccountId: any;
    customerOrderNo: any;
    customerTransactionId: any;
    cuttingApprovedDate: any;
    deletedAt: any;
    deletedBy: any;
    deliveryDate: any;
    discountAmount: any;
    discountType: any;
    documentNo: any;
    employeeId: any;
    eximCreditContractId: any;
    expense: any;
    factoryId: any;
    fine: any;
    forexId: any;
    forexUnitPrice: any;
    generalExpensePercent: any;
    gpsXcoordinate: any;
    gpsYcoordinate: any;
    handoverDate: any;
    insertedAt: any;
    insertedBy: any;
    invoiceId: any;
    inventoryId: any;
    inventoryUnitId: any;
    isAmended: any;
    isChecked: any;
    isClosed: any;
    isCancelled: any;
    isDeleted: any;
    isLocked: any;
    isMultipack: any;
    isPlm: any;
    isPrinted: any;
    isRequirementLocked: any;
    isSample: any;
    isSeparateDeclaration: any;
    isSet: any;
    isTrequirementLocked: any;
    isVirtual: any;
    isWarp: any;
    labRecipeId: any;
    lengthShrinkage: any;
    lockedAt: any;
    lockedBy: any;
    manufacturingTypeId: any;
    orderItemId: any;
    parentId: any;
    packageQuantity: any;
    planDate: any;
    processId: any;
    productionCertificates: any;
    productionColorCumulativeType: any;
    productionCumulativeType: any;
    productionType: any;
    projectId: any;
    proxyNo: any;
    purchasePriceType: any;
    quantity: any;
    recId: any;
    workOrderNo: any;
    recipeId: any;
    recipeQuantity: any;
    repairType: any;
    resourceId: any;
    reviewOrderType: any;
    roadNo: any;
    rowColor: any;
    routeId: any;
    routeItemId: any;
    service: any;



}